Add `minSize` option which adds to `other` result
An implementation of a `minSize` option (proposed in #32).

This allows filtering the output to only show languages with a resulting size greater than what the user inputs.

This is done by moving the results of these too-small languages to an `other` key and deleting the original results.

This may be too ad-hoc - for instance, maybe this should be a CLI-only option and the actual data isn't changed? Is the `other` key needed, or not?

For this reason this feature is not being committed straight to main and will sit as a pull request for proposal.